+++ lisp/eab-org.el
@@ -158,10 +158,10 @@
     "* %?              :NOTE:\n  %u\n  %a\n  %i")
    ("w" "org-protocol" plain
     (here)
-    "  - %t synopsis\n    - %L\n      - %:description" :immediate-finish 1)
+    "  - %t {{{synopsis}}}\n    - %L\n      - %:description" :immediate-finish 1)
    ("s" "org-protocol" plain
     (here)
-    "  - %t synopsis\n    - %L\n      - %:description\n      - %i" :immediate-finish 1)))
+    "  - %t {{{synopsis}}}\n    - %L\n      - %:description\n      - %i" :immediate-finish 1)))
 
 ;; Fix (current-buffer) *server*
 (defadvice org-protocol-capture (before eab-org-protocol-capture activate)